refactor: finalize it of last two commits
