+++ invenio_campusonline/records/config.py
@@ -6,16 +6,16 @@
 # modify it under the terms of the MIT License; see LICENSE file for more
 # details.
 
-"""Services configs."""
+"""API."""
 
 from dataclasses import dataclass
-from typing import ClassVar
 
-from ..records import CampusOnlineAPI, CampusOnlineRESTConfig
+from ..types import URL, CampusOnlineToken
 
 
 @dataclass
-class CampusOnlineRESTServiceConfig(CampusOnlineRESTConfig):
-    """Campusonline REST service config."""
+class CampusOnlineRESTConfig:
+    """Campus online rest config"""
 
-    api_cls: ClassVar = CampusOnlineAPI
+    endpoint: URL = ""
+    token: CampusOnlineToken = ""
